[CI:DOCS]libpod endpoints
small refactor for libpod inclusion and began adding endpoints.
