+++ pkg/api/handlers/generic/system.go
@@ -1,13 +1,14 @@
-package handlers
+package generic
 
 import (
+	"github.com/containers/libpod/pkg/api/handlers"
 	"net/http"
 
 	docker "github.com/docker/docker/api/types"
 )
 
 func GetDiskUsage(w http.ResponseWriter, r *http.Request) {
-	WriteResponse(w, http.StatusOK, DiskUsage{docker.DiskUsage{
+	handlers.WriteResponse(w, http.StatusOK, handlers.DiskUsage{DiskUsage: docker.DiskUsage{
 		LayersSize: 0,
 		Images:     nil,
 		Containers: nil,
