+++ pkg/api/server/register_info.go
@@ -1,12 +1,12 @@
 package server
 
 import (
-	"github.com/containers/libpod/pkg/api/handlers"
+	"github.com/containers/libpod/pkg/api/handlers/generic"
 	"github.com/gorilla/mux"
 )
 
 func (s *APIServer) registerInfoHandlers(r *mux.Router) error {
-	// swagger:operation GET /info libpod getInfo
+	// swagger:operation GET /info ?libpod getInfo
 	//
 	// Returns information on the system and libpod configuration
 	//
@@ -22,6 +22,6 @@ func (s *APIServer) registerInfoHandlers(r *mux.Router) error {
 	//   	description: unexpected error
 	//		schema:
 	//	      "$ref": "#/types/ErrorModel"
-	r.Handle(VersionedPath("/info"), APIHandler(s.Context, handlers.GetInfo)).Methods("GET")
+	r.Handle(VersionedPath("/info"), APIHandler(s.Context, generic.GetInfo)).Methods("GET")
 	return nil
 }
