+++ pkg/api/server/register_ping.go
@@ -1,12 +1,12 @@
 package server
 
 import (
-	"github.com/containers/libpod/pkg/api/handlers"
+	"github.com/containers/libpod/pkg/api/handlers/generic"
 	"github.com/gorilla/mux"
 )
 
 func (s *APIServer) registerPingHandlers(r *mux.Router) error {
-	r.Handle("/_ping", APIHandler(s.Context, handlers.PingGET)).Methods("GET")
-	r.Handle("/_ping", APIHandler(s.Context, handlers.PingHEAD)).Methods("HEAD")
+	r.Handle("/_ping", APIHandler(s.Context, generic.PingGET)).Methods("GET")
+	r.Handle("/_ping", APIHandler(s.Context, generic.PingHEAD)).Methods("HEAD")
 	return nil
 }
