+++ pkg/api/server/register_system.go
@@ -1,11 +1,11 @@
 package server
 
 import (
-	"github.com/containers/libpod/pkg/api/handlers"
+	"github.com/containers/libpod/pkg/api/handlers/generic"
 	"github.com/gorilla/mux"
 )
 
 func (s *APIServer) registerSystemHandlers(r *mux.Router) error {
-	r.Handle(VersionedPath("/system/df"), APIHandler(s.Context, handlers.GetDiskUsage))
+	r.Handle(VersionedPath("/system/df"), APIHandler(s.Context, generic.GetDiskUsage))
 	return nil
 }
