+++ pkg/api/server/register_version.go
@@ -1,12 +1,12 @@
 package server
 
 import (
-	"github.com/containers/libpod/pkg/api/handlers"
+	"github.com/containers/libpod/pkg/api/handlers/generic"
 	"github.com/gorilla/mux"
 )
 
 func (s *APIServer) registerVersionHandlers(r *mux.Router) error {
-	r.Handle("/version", APIHandler(s.Context, handlers.VersionHandler))
-	r.Handle(VersionedPath("/version"), APIHandler(s.Context, handlers.VersionHandler))
+	r.Handle("/version", APIHandler(s.Context, generic.VersionHandler))
+	r.Handle(VersionedPath("/version"), APIHandler(s.Context, generic.VersionHandler))
 	return nil
 }
